Reupload check performed at row level
The reupload chaeck is used to work out if substitution
is required on row updates. Perviously it used manifest
state, but this only matched the state of the most recent
upload, thus failed was ambiguous.
- Failed first upload
- Failed reupload
This change looks directly at the state of the external library
creation request, which directly reports on whether the sample as
been moved downstream.

TODO: We don't currently support manifest reuploads for plain
libraries. These could use a similar row level mechanism, although
the actual check itself would need to be different. (Probably just
counting the number of aliquots with the same library_id, otherwise
we'd need to cover all the ways in which a sample can exist elsewhere)