Switch label sent to printer
Ensures that the label content remains consistent, even if old labels
are re-printed.

Progress note: Discussion with JL reveals that WorkingDilution plates
aren't typically loaded on the Tecan, and a quick database query
revealed that this was the case over the affected perios. (Last case was
three years ago) This indicates that I may have misunderstood the
problem.