Fix manifest substitution issues
Tag substitution would only screen clashes in the tags which had
been substituted. Which meant that in non-udi dual index pools
there was a risk of falsely detecting a clash if only the i7 tag
(or the i5) was substituted. Given we're already detecting clashes
within the manifest, AND, have fallback tag clash protections in the
database, we just disable this redundant round of clash detection.

This bug also revealed that we were improperly propagating errors
back out to the upload. (My fault) This ensures errors get properly
propagated.