Use find_by! to raise error if id missing
Some of our power users modify the url to access reports directly, however if they typo, we get a method missing, as the find_by returns nil. ::find_by! will raise an ActiveRecord::RecordNotFound instead, which gets handled automatically, and converted to a 404. Not only is this more informative for the user, but it stops us getting spammed.