Deprecate manifest api endpoints
The endpoints have been unused for over two years. I have confirmed
with the last study using the API that they no longer need the
functionality.
Other studies have not used the API for a substantial length of time.

While it would be possible to make some modifications to the API to
provide only minor changes to the interface, it would prevent the
removal of significant chunks of code, and be counter to the general
drive towards migrating to the v2 api.