SEQ812 - Refactor sample/study lookup
- QcResults have studies and samples through assets
- Asset association specifies receptacle as non-receptacles don't have
samples in the same way so the has_many through fails

I'm not too certain this is actually the correct move. While it
does avoid the need for the qc assay broadcaster to know anything
about how qc_results work, the number of adjustments to support
it make it feel like the naieve approach is probably simpler.

Main difference is it means we can't associate a QC result with
an plate, only wells or tubes. This actually makes sense in the
current context, and may even need to be enforced if we ever fix
the labware/receptacles mix.