ss773 - Add optional tags attribute to submission
Note: jsonapi-resources does not currently support
attributes that are not rendered by default, and assumes that
is a sparce fieldset is not specified, that all fields are required.

There is currently an issue
https://github.com/cerebris/jsonapi-resources/issues/855 to suggest the
addition of this functionality.

This functionality would be failry easy to add, it would involve
modifications to: JSONAPI::ResouceSerializer.supplying_attribute_fields
to add the filtering, as well as extension of JSONAPI::Resouce to allow
the configuration of default attributes.