Improve work completion performance
Work completions were timing out, especially in the GBS pipeline

N+1 query problems:
1) Only check primer panel on create
2) Eager load Order upfronti

This only shaved off about 10% so ruby-prof was used to identify
bottlenecks

1) Apply_library_information_on_pass was slow. This was mostly due
to slow performance of the rails mutation tracking. However, as the
method has been redundant for almost a year, it is probably safe to
remove. IN the unlikely event that older work does get restarted
it is trivial to repair.

2) Oddly accepts_nested_attributes_for was causing validation
to be run on all target_assets, and then also on their aliquots.
This happened even if target assets were being assigned directly.
Given we don't use this functionality (as best I can tell) I have
removed it.