Allow renaming of studies
- Add a study name field to the manage study page
- Validate uniqueness of study name on update and create

Note: This renders 22 Studies in the Sequencescape database invalid
as the have non-unique names. THis will need to be addressed
before this story can go live. SSRs have been emailed the following
(Study names and ids removed)

Hello all,

I’m currently adding the ability to rename studies via the ‘manage’ page in Sequencescape. As part of this I have discovered a number of studies which share identical names:

<snipped>List of ids and names of duplicate studies

I suspect many of these have been created in error, and will definitely be very difficult to work with in Sequencescape. Once we’ve added the ability to rename studies, you will also observe further issues, as Sequencescape will see the studies as invalid.

For each set of studies we have the following options:

1) Attempt to merge them together. In this case, please let us know which study we should merge things into.
	- We’ll keep the old study around, but add ‘(Legacy)’ to the end, a comment on the study will record what happened.
2) Differentiate them by renaming one or both studies. (Eg. Adding a year)
3) Abandon the plans to let you rename studies.

Leaving them unchanged , without taking option 3, is unfortunately not an option. If the studies are invalid it will end up having an impact on everything associated with them, as validation issues tend to cascade and affect anything associated with them, even sometimes indirectly. (eg. Submissions, samples, requests, plates, study reports, possibly even lanes associated with the study)