Bug: Aliquot indexer tags all aliquots
Sample manifests were creating aliquots with only an i5 tag. While this
behaviour isn't actually required in the long term, it resulted in the
generation of lanes with no indicies. Subsequent repair of the data
would not trigger re-indexing.

There is little reason to make the aliquot indexer conditional so this
has been removed to improve the robustness of the process.