Remove PhiX checks
In most cases PhiX is associated with the lane by setting a SpikedBuffer
tube as a parent. Both the batch.xml and Flowcell::IO use this
relationship to find the PhiX. I'm not entirely sure WHY things are set
up this way, it may be that PhiX predates aliquots and multiplexing.

The code in question deals with a hadful of exceptions in the database
where PhiX aliquouts were present in the Lane. This prevents the PhiX
from being misidentified as an actual sample. Alsongside this work I
have removed the excess PhiX aliquots, preserving the standard
relationship. This makes this code redundant, but doesn't affect the
output of batch.xml or FlowcellIO.