Fix aliquot remover
The aliquot remover was wrapping destroyed aliquots in a custom class
under the mistaken belief that we could not longer retrieve information
about destroyed aliquots. This was failing as it didn't impliment the
full suite of tag methods.

However, this class was entirely unecessary, so has been removed, and
the destroyed aliquots are used directly.