Fix PacBio inbox
This fix was performed in the imported changes, but was missed in the
initial import as it had occured in a different commit.

Previously the item id was being used to group requests. with the
removal of item from Submission creation this was failing. The initial
fix generated a uuid, but the index already provides a unidue index, so
we use that instead.