Switch activeresource to gem
Active resource was using the master branch on github.
This was a result of changes which had not been properly released at the
time of the Rails 5 switch.

However the latest release contains all necessary changes, so we switch
to using it to reduce risk and confine ourselves to official releases
only.