Remove Event::AssetDescriptorUpdateEvent
This callback had NEVER sucesfully fired in production, as the
descriptor key was getting set on an update, after the after_create
callback had fired. As it will no longer work, and the routes which
might have been expected to set it have been removed anyway, we just
remove it.