Remove study relation functionality
This functionality has been in place for several years and has never
been used. THis commit removes the associated behaviour. A later commit
can drop the tables study_relations and study_relation_types. I'm
keeping them in place for the moment to avoid data loss in the event
someone uses the feature between now and this commit being applied
in production.