Load SimpleCov before linting factories
Code coverage only flags a line the first time it is run, so it
is important that SimpleCov is loaded very early on. If we lint
before loading, then we under-report coverage significantly as any
code hit by linting will have zero coverage, EVEN if it is subsequently
hit by our tests.