GPL060 - Add receptacles/:receptacle_id/parent
The asset refactor will eventually impact the baehaviour of assets.xml,
with the long term goal being to remove this page entirely. Research
through Splunk revealed a single user, who was scraping the page to
retrieve the multiplexed library tube id for lanes. This was in order to
provide a deep-link for users of their own software.

When applied, this commit adds a new endpoint 'receptacles/:id/parent'
which will automatically direct users looking up a lane to the
corresponding multiplexed library tube. This removes the need to scrape
assets.xml, and drastically minimizes shanges both in our software, and
that of the user.

The endpoint also aliases the assets#show as receptacles#show to reduce
friction as future changes are made.

Given the endpoint may also recieve non lane receptacles we also handle
the following:

No parent: 404 page linking back to the receptacle
Multiple page: 300 page listing all parents

Parents are found via requests as:
1) Asset links tend to link labware, not receptacles
2) Transfer requests aren't actually present for lanes, just SequencingRequests
3) Requests are more customer focussesd