Fixes #2216 - Samples not generated up-front
Bit of a roundabout route to get here:
1 - Samples were getting generated the first time they were accessed
2 - This was happening in the initializtion firstly where their
existance was checked, and secondly where they were passed in to the
specialized fields
3 - Updating the fields to accept the sample_manifest_asset instead

Then it was mostly a case of fixing up the tests, ans making sure
everything gets routed through the same instance of 'sample'. This
invloved migrating the sample generation onto sample_manifest_asset.

Both old and new tests are not passing, so committing prior to the
refactor. Next steps are:

1) Make sample generation explicit prior to sample update. SHould reduce
the risk of the problem occuring in future.
2) Do a rubocop pass on the files changed
3) Explore caching the sample manifest affests and their samples etc.
upfront, which should drastically improve performance.