Add webdrivers to manage chrome-driver
Our travis ci builds are breaking due to a mismatch between the chrome
version and chromedriver. This is mainly are there doesn't seem to be a
single source which provides both, and is similar to an issue we had a
couple of months bak.

While chrome-driver does seem to install a compatible version of
chromium, our suite uses the installed chrome version that appears to be
par tof the base image. I briefly tried specifying this with
CHROME_BINARY but had no luck.

The weddrivers gem automatically matches the webdriver version to the
chrome version. If we don't hav emuch luck here we'll probably need to
be installing chromedriver manually.