Add character validations to warehoused fields
THe warehouse is utf8 format not utf8mb4, which means it doesn't support
four byte supplementary characters, which include most emoji 😭 as well
as other elements, like rare kanji. We add validation to columns which
get sent to the warehouse (and aren't otherwise restricted) to ensure
that the messages dont' get rejected there.

We also apply some additional constraints on selected columns to relieve
concerns from NPG that our database changes will result in the
propagation of more special characters downstream, that some of their
pipleine may be unable to handle. These restrictions may be adjusted
over time to better reflect the actual limitations, and to ensure that
all relevant fields are covered.