Fix null delay permission blocking update
On updating studies the hidden data_release_delay_approval field
wasn't being properly disabled. For update studies without a
delay this value was null, and so the form was rejected as having
incomplete fields. However because they were hidden, the browser could
not show the error, and just did nothing.

The related fields javascript is supposed to handle this, however it
interacted badly with the nested options, so got disabled, then
immediately enabled again.

While this bug has existed for a while, it was previously hidden, as the
select dropdown meant it was defaulting to 'No'.

This fix moves the field out, to be visible for all delayed studies, not
just PhD sttudies. Spoke to RC and LE, who agreeed. Originally the field
had been hidden for PhD studies, as they are approved by default.
However, as a result of this bug, they were inadvertantly being recorded
as unapproved anyway.

Following this change the SSRs WILL need to select *yes* for PhD
studies, but at least the correct information will be recorded.