Improved robustness of aliquot equivalence check
The previous version of the check compared hard-coded
attributes when checking equivalence. This resulted in
newer columns, such as primer_panel_id getting missed
from the check.

To avoid the need to remember that this method needs to
be updated when extending Aliquot, the columns to compare
are now generated dynamically from the database columns.
Columns which are permitted to change are now whitelisted.

This should substantially reduce the risk of future bugs
being introduced, and will ensure that the comparison
remains conservative, rather than permissive.