Fix validation of study_metadata on accession
Changes in Rails 5.2.4 https://github.com/rails/rails/issues/28080
changed the way validations of associations were performed. This
prevented it from our guards triggering validation on accessioning
as Rails was not detecting the record as having changed.

However this process was always stretching the Rails validations,
and was hard to follow. Fortunately, in Rails 4.1 they added custom
validation contexts. Which is exactly what we need, and MUCH cleaner.