Switch to Ruby-2.6.6 and compatibility fixes
Ruby 2.7 was heavy on the deprecation notices, mostly related to the
changes in named parameters. Although general compatibility appeared
to be the same as with 2.6. Given the final changes to named parameters
have been delayed, it migh still be worth considering this switch,
fixing our code, and supressing the warnings.

The only failing test was related to changes in the CSV parser.
PReviously we relied on the file IO being read in one line at a time by
the CSV parser, so that we could resume reading the file once we'd found
the header location.

However, in Ruby 2.6 this can no-longer be relied upon, as the CSV
parse reads in more of the file at a time. As a result, gets on the CSV
and gets on the file aren't synchronised.

This change is a little hacky, but I wanted to maintain the header
functionality of the CSV libraries. Instead before reading in the file
for the CSV body, we rewind it, and then step forward until we reach the
header row.