Separate control plates for Hamilton
This updates the expected_layout for the Hamilton robots.
Unfortunately its unclear how we'll hook
this into the worksheet generation as
we don't persist the robot selected during
cherypick processing.
We'll also need to ensure we create the right
robot verification class when we do the verification itself, so will
probably need to extend either the Robot
class, or its properties.