Fix intermittent spec failure
Asset_path disambiguation will handle any cases where there is
not a clash in labware and receptacle barcodes. However, if there
is a clash, this page will not redirect and will ask the user
for confirmation. This happens more often if the test falls
earlier in the test suite.

This fix uses the correct path, labware_path, rather than the
asset_path