Filter all requests if none selected
If no requests were selected, no condition would be applied.
Instead all requests on base scope would be returned.
By returning false if there are no conditions we ensure that no requests
will be found.

We also add a bit of error handling to the requests controller to prevent creation
of empty batches. Ideally that would be handled
by the batch validation but things
are currently hooked up incorrectly for
that, and this fix is urgent.