PickList requires submission
- Enforce pick-list submission requirement.
- Autosave also ensure that submissions are valid

Autosave true may be slight overkill here, but it ensures:
- Validation of the submissions is properly handled
- Error messages are propagated back to the pcik-list

Relying on the standard autosaving of newly added associations
reveals unexpected behaviour. The submission
fails validation, so isn't saved, but the pick-list
goes ahead and gets saved anyway!
This results in the database constraint triggering,
preventing the save in a very messy manner.

validate_associated solves the above behaviour, but results in the
somewhat useless 'submission was invalid' error.

We do have validates_nested, which essentially is a home-built
implimentation of the same behaviour as we get with autosave.
Might consider that later, but keeping things Rails native for now.