Try resetting barcode counter before each test
The barcode number sequence ensures we maintain unique barcodes.
Without this line it it scoped to the test suite, which means that
test barcodes vary depending on when in the test suite the test
is run. While this avoids unintentional coupling of test
expectations to the order in which plates are created, it results
in frustrating random failures if the generated barcodes happen
to coincide with explicitly defined ones.