Fix directory creation race condition
Saw this on Travis. Assuming chrome was creating
the directory between us checking if it existed and us actually
creating it. This will rescue the condition, and retry.

Retrying may seem odd, but we might have failed to create
the parent directory, and still need the downloads directory.
The new PATH.exist? check will stop us getting in an infinate loop.
Although it is possible that we'll race on the parent path, and
then race again on the child. But on the third time through we'll
be fine.

Lets just hope that chrome doesn't crash if it loses the race!