Prevent job being wrapped in a transaction
By default aasm (since 3.0.13) wraps state changes in a transaction.
In most cases transactions are a good thing. However here we had
specifically chunked the job into batches, each with its own
transaction. This was to ensure that reports over 1000 samples in
size would require a consistent level of memory.

While we still retrieved our records in batches, Rails holds
on to any records with an 'after_commit' callback for the duration
of the transaction. This would include any aliquots for which a
fail descision had been made.

By removing the outermost transactions, we broadcast the aliquot
updates in batches.

Although on writing this commit it occurs to me that the aliquot
is fetched as read-only. Might need to fix that.