Exclude pending batches from returned plates
Pending batches don't have pick information, so result in a error
message on the right hand side. This was never perticularly elegantly
handled, as the plates would be stuck waiting for the batch information.

During UAT however, the uses expressed that they'd prefer to hide
incomplete batches in their entirety.

This change ensures plates should only include the list of batches
for which they expect to have pick information.