Document edge-case implications
Following a discussion with jgrg today it became apparant that we have
a potential future risk should we change our handling of cross-study or
cross-project orders. While it would not be *incorrect* to associate
multiple studyes with an order event, it would reduce utility for users
of the event warehouse.