Add github action for Rubocop Linting
Travis change in pricing has meant we need to look at alternatives.
While CircleCI is already in place as an emergency fallback, its free
tier doesn't quite provide the resources of the previous Travis setup.

This action:
1) Sets up a VM with Ruby 2.7
2) Installs dependencies required for linting
3) Sets up caching of gems and cops for improved performance
4) Runs the linting

It does not:
1) Update Rubocop to a version that uses the RUBOCOP_CACHE_ROOT
   environmental variable
2) Make use of advanced inline reporting features.

The latter allows failures to be reported back as inline comments on the
PR. This looks cool, but apparently causes some issues with pull
requests from other repos:

A deprecated action that added all that:
https://github.com/andrewmcodes/rubocop-linter-action

Refine lint