Update tests to check behaviour
The previous tests were more closely linked to implementation
than behaviour. This avoids testing specific values, and instead
ensures the data meet the criteria we need.

NOTE: I'm a little uncomfortable with the distribution of the
random data. As a single well being hit 25 time is surprising.
If I'm dredging the correct statistic out of the bottom of my
brain (and using online calculators correctly)

1000 trials with a probability of success of 1/96 (0.01041666) means
we can expect to hit the same well >= 26 times with a probability of
0.00003 . Now, that's mitigated by the fact we're actually performing
96 similar tests at the same time, but its still a low probability.

We don't see the same issue for other batch ranges, so it doesn't
seem to be systematic. Just probability being annoying.