Drastically prune back library
We barely use any of the functionality ot the legacy
plugin, so we strip it back to the bare essentials here.

In practice it just comes down to a handful of is_<role>[_of]
method_missing handlers and the has_role method. (And the recently
patched has_no_role) As well as some of the rails associations.

This is all trivial stuff to re-implement in basic rails.