Add documentation about test failure
When I added the deprecation for the sample reception, I didn't consider
that tests would begin failing once the target date was reached. This
caused some unexpected failures with an unrelated update, and delayed
an important release. Added a note to the deprecation card library to
reduce the risk of a similar error being made in future.