Add documentation and adjust styles
- Add documentation to the controller
- Make the hide show comments links more prominent

I had hoped to improve the comment presentation, but this is a little
tricky given the table structure. Ideally comments would span the
whole width of the table. This could be achieved by displaying them
as separate rows, but things get a bit tricky when it come to
compatibility with multiple inboxes, and variable colspans.