Remove PlatePurpose::InitialPurpose behaviour
This plate purpose is only actively used in the PCR free pipeline
and where its use of this subclass is an accident of migrating
pre-existing pipelines.

The custom behaviour was solely concerned with looking up the
pending orders for broadcasting library start messages.

There is no reason to suspect that the default behaviour is
insufficent.

I'll double check everything with the integration suite, but
no major worries