Fix submission creation privileges
Fixes #3112

Creation of a submission relies on the order_fields and study_assets
routes on the SubmissionController. However as we use authorize_resource
to provide authentication this requires these be granted as specific permissions.

These permissions are missing from all roles other than administrator, which
grants them via the wildcard manage permission.

This fix adds an alias to ensure that the ability to create a submission
automatically grants all other required privileges