Remove aliquots delegation to receptacle
The delegation to receptacle handled building of aliquots on single
receptacle labware (eg. tubes) but prevented the eager loading of
aliquots and their information via the API. This tanked performance for
tubes.

Fortunately we only really relied on the ability to do this in a handful
of places, the majority of which were in the factories. Manually fixed
these up to go explicitly via receptacle.

We do keep the delegation in place for aliquots= as there are currently
a few too many places using that for comfort.

It would be nice if rails let you provide hints about constructing
records via has many relationships.