Extensive test coverage of state changers
Some comments in the review indicate that it is probably
a good idea to try and rationalise the slight differences in
logic driving update of outer state.

Generated a whole load of tests to cover current behaviour, will
probably reduce them down to core behaviours after I've made the
changes.