Add config option to drive default behaviour
My suspicion is that we'll need some more complicated behaviour
than this eventually, with the 'free' wells varying between processes.
I contemplated adding the configuration to the PlateTemplate (on plate
metadata) or the plate purpose (a new field?) but am guessing at
future requirements, so kept it simple instead.

I avoided passing the options through from the front-end, as its
just another point of failure when it comes to user-error.