Remove ... to allow flog to work
Flog was chocking on parsing ... for argument forwarding
but given we're only using this on ine place, where it was
entirely redundant, we can loose this.

However...

I'm not sure flog is hugely useful, as I can't find out
how to configure it to ignore RSpec DSLs, so seem to get a lot
of false positives. Plus, Rubocop already reports on ABC
complexity.

So will fix it, but tempted to remove it completely.