Remove flog and flay from dev dependencies
Flog and flay are code metric tools which currently run as part of
code climate, in addition to having much of their functionality covered
by rubocop.

We're not in the habit of running them routinely (flog has been broken
for years) and they introduce several downstream dependencies. In
addition, the gems don't need to be bundled to be used.

As a result, I'm removing them.