Change queue for ease of deployment
Our current queues are inconsistently configured
with different dead-letter configurations between
environments. In addition, many of these have no
queues bound.

If a pre-existing queue is declared, its configuration
needs to match. Here we change the queue name to
let us avoid clashing with existing configuration.

This will require that the old queues get removed,
but it lets us manage this in a more controlled manner.