Fix multiple accession jobs
Accession jobs were generated each time the sample was saved during manifest processing.
This commit adds tests for the extra
jobs, and removes the extra save calls.

I'm going to see if I can add a belt
to these braces, as its the kind of
error that is easy to slip back in.
(And wouldn't get picked up by tests
if the column wasn't in the tested
templates)