Fix key encoding problem
We mostly saw this as + being lost from the begining of keys.
Furthermore, attempts to repair the data resulted in the same error.

The culprit was an attempt to remove
escape characters on validation,
which was triggering twice, removing the +.

However, all this was handled already, and was only added here due to
unecessary escaping of field names in the HTML template.

All fixed!