Remove pdfmake dependency
PDF make provides an extension to data-tables to allow the export of
pdf files. However, we aren't using this functionality. I'm removing
pdf-make because it introduces a large number of dependencies, some
of which have security vulnerabilities. It does look like there is some
work to move some of these onto more actively maintained forks, but
my quick attempt at an update suggested that some of the transient
dependencies still require the unmaintained libraries.