Fix frozen string helper
We actually only use this helper on a single page, so I half considered
removing it entirely. However, I feel it would make more sense to
migrate other select fields to using it as it reduces the propagation
of bootstrap classes through the code.

I also updated the other class appending fields to use the same process,
as otherwise they would break if a frozen string were ever passed in.