Extract limber tube and plate purposes to record loader
Its not ideal having plate and tube purpose duplicated in both
sequencescape and limber, as it can results in conflicts between
the two representations. As a result, I've treated production as
the authorative source when building this list. As a result:

1) This behaviour is slightly different from the previous seeds
   mostly in terms of seeting barcode printer type.
2) Not all types are as uniform as I'd like, althoug the differnces
  alre almost entirely inconsequential.

I'd like to address point 2 by making to production data consistent,
but am not about to do this before I run off on a two week holiday.

In addition, the limber task had accumulated a handful of purposes
that were not limber related. I've grouped these in their own files.