Switch real-time sortable AJAX updates for manual
Fixes #3253

Automatically updating batch location on each table update
was resulting in race conditions if multiple requests were handled
at the same time. Occasionally we'd also end up with an inconsistent
state, which given the presence of transactions I'm assuming was a
result of invalid payloads.

This changes switches the previous updates to an explicit save, and
switches all data handling to be conducted via data-tables.