Fix missing button styles
The upgrade to Sassc broke some button styling. It turns out that
this rule was the problem, and was resolving to a transparent
background colour.

It appears that this was a side effect of the not(.btn) rule clashing
with the subsequent import of btn.

Fortunately this exclusion was entirely unnecessary, as the usual
css rules mean that classes count as more specific operators, and
override the default button styles anyway.