Fix tests failing due to validations
We now validate uniqueness of library type name
But lots of tests were violating this constraint.
Mostly its actually been a case of just not associating
library types with request types.

Unfortunately I have a test-order dependent failure in tests involving
the DownloadHelper. I think its unrelated but need to get it working
as it makes it impossible to run tests locally.