Add INSDC countries
The INSDC provide a controlled vocabulary
for describing countries. This is required by
the ENA when describing the geographic
location in which a sample was obtained.

This data will shortly be required:
https://github.com/sanger/sequencescape/issues/3470

Currently however we have a free text field,
which is resulting in some data-integrity
issues.

In order to improve integrity in future we
are tnormalizing the data against this table.