Switch to caxlsx
Fixes #2584

This one proved simple. I had been concerned by the v2 -> v3 version
jump, but the changelog was reassuming, and it was mostly dropping
support for earlier versions of ruby. A few other breaking changes, but
nothing major, or anything we used.

As it was, pretty much a drop in replacement. Wish we'd done it earlier.