Fix updating of library_type
Previously the use of the name column worked,
but it caused the standard library_type column definition
to become over-written. This adds the ability to define
an update instead.

This essentially makes the need to customize the
name meaningless, so it can be removed. This will follow
in the next commit.