Migrate stylesheets to vite
The hardest bit of this was setting up styles previously provided by
gems.

In the case of fontawesome I just switched to the npm package. This
took longer than expected, as it took a while to work out I needed both
the regular and solid css imports. Spent ages assuming it was an issue
with importing the fonts themselves.

Formtastic was surprisingly easy in the end. I can pull the path out
of Gem and then pass this through to vite. Makes it pretty simple to
reference the plain css files. It did mean I needed to inspect the
path within the gem, so we might see it break again in future if this
updates, but fortunately this is a very noisy failure.