Exclude checking for receive over have_received in the test file
When using receive, the test expects the very next call to that spy to be the one that matches.  have_received allows the call to be later than the next one.  Logging matchers are always chaotic so the test is better when it's able to pick up on any call to the log messages.