Define YAML column permitted classes as strings instead of classes
This resolves in issue where this list wasn't being set properly because it had to be wrapped in an after_initialize block.  My assumption is that we were setting this too late and the app wasn't accepting the values.  We were also bypassing the config object which was potentially going to cause further headaches down the road as it's undocumented.